refactor: separate NavLink to its own file
